Add exec/2: posix_spawn
This is an initial MVP with quite a few things still missing (such as:
better error messages, documentation, tests).
Despite this, it is already feature-complete on POSIX platforms (on
Windows it currently reports an "unsupported on this platform" error).

The signature is `anything | exec(path; [args…])`.
`path` and all `args` must be strings.
`anything` will be converted to a string if it isn't a string in memory,
then piped into the process' stdin.
The output is all stdout of the process.
The exit code is not reported.

Technically, "path" can be a simple name and `$PATH` will be searched.
This is because the underlying function is `posix_spawnp`.
This can bec hanged easily.

The process does not have access to environment variables.
This can be changed as well.

Piping between programs works. Here's an example to try it out:
`tostring | exec("seq"; [.]) | exec("wc"; "-l")`
Expected output when inputting numbers is that number,
but it notably goes through seq, then line-counting.